factor out helpers
